% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_big_epa_cars.R
\docType{data}
\name{big_epa_cars}
\alias{big_epa_cars}
\title{Vehicle fuel economy data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 41804 rows and 83 columns.
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2019/2019-10-15}
}
\usage{
big_epa_cars
}
\description{
From fueleconomy.gov via Tidy Tuesday project
}
\keyword{datasets}
