% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\docType{data}
\name{plan}
\alias{plan}
\title{Analysis plan}
\format{
An object of class \code{drake_plan} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 3 columns.
}
\usage{
plan
}
\description{
Analysis is rebuilt using \code{drake::make(plan)}
}
\keyword{datasets}
